+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.2.11] - 2020-11-17
+
 ### Changed
 
 - `findSlaWithChannel` now returns the cheapest SLA.